Remove forbid_global_names alias
