+++ src/pure_function_decorators/__init__.py
@@ -1,14 +1,12 @@
 """Public package surface for the pure-function-decorators project."""
 
 from .enforce_deterministic import enforce_deterministic
-from .forbid_global_names import forbid_global_names
 from .forbid_globals import forbid_globals
 from .forbid_side_effects import forbid_side_effects
 from .immutable_arguments import immutable_arguments
 
 __all__ = [
     "enforce_deterministic",
-    "forbid_global_names",
     "forbid_globals",
     "forbid_side_effects",
     "immutable_arguments",
